--- client/src/components/related_items/mockData/mockData.js
@@ -1,25 +1,0 @@
-import React from "react";
-const data = [
-    {
-        category: 'plumbing',
-        name: 'pipe',
-        price: 200,
-        rating: 4.5
-    },
-    {
-        category: 'kitchen',
-        name: 'spoon',
-        price: 10,
-        rating: 5
-
-    },
-    {
-        category: 'furniture',
-        name: 'couch',
-        price: 500,
-        rating: 4.5
-    }
-];
-
-export default data
-
